pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "arcagi2"
version = "0.1.0"
description = "ARC-AGI 2 submission"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic==0.75.0",
    "daytona==0.128.1",
    "ipybox==0.6.2",
    "jsonschema==4.25.1",
    "openai==2.14.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "pandas==2.3.3",
]

[project.scripts]
arcagi2-solve = "arcagi2.solver.solver:main_cli"
arcagi2-submit = "arcagi2.evaluation.evaluate:main_cli"
arcagi2-score = "arcagi2.evaluation.score:main_cli"

[tool.setuptools.packages.find]
where = ["src/"]
include = ["arcagi2*"]
